requirements.txt
Django==2.0
django-cors-headers==2.4.0
django-enumchoicefield==1.0.2
django-filter==2.1.0
django-rest-framework==0.1.0
djangorestframework==3.9.1
djangorestframework-simplejwt==4.3.0
Markdown==3.0.1
PyJWT==1.7.1
pytz==2019.2
